Add specific WiFi firmware missing warning and stop start crashing
Also debloat viceapp and move network helper files to it's own file
