+++ Makefile
@@ -19,7 +19,7 @@ endif
 EXTRAINCLUDE += $(APP_INCLUDES)
 
 OBJS	= main.o kernel.o new_io.o vicesound.o vicesoundbasedevice.o bmcmodem.o \
-		  viceoptions.o viceapp.o network_time_sync.o fbl.o crt_pi_idx.o crt_pi_rgb.o
+		  viceoptions.o viceapp.o vice_network.o network_time_sync.o fbl.o crt_pi_idx.o crt_pi_rgb.o
 
 ifeq ($(MACHINE_CLASS),RASPI_PLUS4EMU)
 OBJS	+= plus4emulatorcore.o
